split out ui module
