+++ src/components/list/product.tsx
@@ -4,7 +4,7 @@ import type { LinkProps } from "next/link";
 import type { PropsWithChildren } from "react";
 import EggBasketIcon from "~icons/custom/egg-basket";
 import MilkIcon from "~icons/custom/milk";
-import { ToolbarLink } from "../Toolbar";
+import { ToolbarLink } from "../ui/toolbar";
 
 type CardProps = {
   image?: string | null;
